moving getsourcelogins to utils
